Interaction changes
Got rid of the way I handled interactions before this. Came up with a new way to handle the movements of the hands, because the previous one was stuttery, and caused a lot of problems. This was caused by the rigidbodies not being able to act smooth with the library. Also worked on the backpack for a bit, and added the function to put a hat on your head and put the backpack on your back. And last but not least I started working on the weapon holsters.