Weapon effects
Added weapon and other sound effects to the game. The PlayerController is finished for now. Every mechanic that I wanted works (particles and sounds added in this commit). I made a soundsystem that is easy to use from every script. The next stop is the AI for the game. After that I want to add an event system, and then we are done with the mechanics!