Started with the AI
Made a first prototype of the AI. The AI chooses a spot to take cover, and shoots at the player. The AI also has accuracy, beacuse it would not be fair if the AI hits every single shot it takes. The AI movement works with a navmesh. This navmesh gives the AI info about where it can and can't walk.