Big Enemy rework
Made a lot of changes to the AI behaviour. Should've splitted this commit up, but I did not (I am a idiot for that). So to round it up as well as I can.
One of the changes is that the enemy now moves and chooses coverspots dynamically. If the player is not in range, the enemy runs towards the player. When the enemy is close enough it chooses a spot behind a cover layer object (that is closest to it).
After taking cover the enemy shoots using the pistol/weapon class. I also made changes to the spawner. The spawner now chooses random positions to spawn an enemy, based on the spawnpositions in the array. It also keeps track of waves. If all the enemies are dead for the wave, the system updates to higher values. The values are kept on track with a event/action that it listens to. The AI calls the action when it dies via the GameEvents class, where all the events are going to be addded.  The event gets called and the EnemySpawner class listens and changes the values.