Added slomo effect
Added a chance for a slomo effect, when an enemy gets killed(50%).  Did this by changing Time values, and also applying the values to the pitch of the sounds. This way everything looks and sounds in slomotion. Implemented the funciton calls with the GameEvents class. This way everything stays loose!