Started working on the level disign
Started working on the level design. Added colliders to objects, added the correct layers to the correct object (mainly cover objects to the cover layer). Also added some post processing, so that the game looks a lot better. Next up I need to add some spawn locations for the enemies / add events that are connected to the player's presence.