Added SpawnEvents and Dialoguesystem
Added SpawnEvent triggers, where something gets spawned based on the triggers values. Also added a dialogue system that prints out text, that you assign to it. This text gets sent to the DialogueManager(The manager handles the sound/and the typing). The manager also checks, if the audio is stopped playing. After it is stopped it plays the next sentence that is sitting in the que. After alle the sentences are said/printed out, the dialogue ui dissapears.  The dialogues can also get triggered via colliders and via script. The script works via the StartDialogue, and the DisplayNextSentence functions.